Move CoverageRangeRecord to separate file, add DebuggerDisplay
